Split flightdeck into snippets, add gradient; change lvds setup to apb.
